Initial commit.
Note that I did not actually create the repository at the time recorded
on this commit (some time in 2009) -- the repository was actually
created at approximately 2013-08-29T15:53:04-04:00 with one huge initial
commit of around 1000 files. However, at approximately
2016-11-14T22:20:00-08:00, I went back and manually reconstructed a
series of smaller commits dating back to 2009, based on the time the
file was last modified on my hard drive (mtime). This allows us to store
a more accurate history of when I actually solved the problems.

Unfortunately, the method is far from perfect. Before using rsync, I
would back up my code with methods that did not preserve mtimes. This
led to a large number of files on my hard drive that have an mtime of
around 2009-10-08T18:33:27-04:00, which was presumably the first time I
started using rsync.

Furthermore, Git commits store time zones, but Unix files do not. Thus,
I had to manually derive (or guess) based on the time zone in which I
updated the file.

Additionally, there are several instances in which I've modified or
moved the file after the original initial commit in 2013. In this case,
the original timestamp was lost. I was able to re-derive the original
timestamps in the cases where the file was moved, but not in the other
cases. Those final remaining files will be committed with the timestamp
of 2013-08-29T15:53:04-04:00 -- when I originally created the
repository.