Google Code Jam 2018 Round 1C
Round 1C was as surprisingly easy as 1B was difficult, but I was rusty
and made a small blunder. As a whole, I simply need to practise more.

A Whole New World:
This was a rather straightforward recursive search problem. I was hung
up on trying to come up with an optimal trie solution, but the slow
solution was sufficient.

    * The fast solution is somewhat counterintuitive because you're
      recursing with a trie, but you should not think of it primarily as
      recursing *along* the trie's edges.

    * The slow solution is less elegant but is still extremely fast
      given given the size of the input. I should just consider the slow
      solution next time.

Lollipop Shop:
My main weakness in this problem was failure to commit. I got the
intuition to sell the least-seen item the earliest, but I couldn't prove
to myself that it would be a 0.9 approximation. There are two ways to
get more certain:

    * Just test the problem. The only set is the visible set anyways, so
      there is zero risk that you'll get the visible set and fail the
      hard set.

    * There is no more information I could possibly get that would help.
      This is a trickier argument but should still be pretty obvious
      given the situation.

Ant Stack:
I had two issues with Ant Stack:

    * I was slow to come up with the longest-increasing subsequence
      solution. I jumped to a constructive solution (how can we build
      all sequences) instead of a deductive one (how can we characterise
      optimal sequences). The second approach would have led to the LIS
      approach. I was also a bit suspicious of how simple this was and
      expected something more difficult.

    * I blundered when writing the LIS dynamic program. Instead of
      updating all possible sequences, I only updated the longest one.
      This is because I was applying a straightforward calque
      translation of the LIS algorithm, which doesn't work in this case.