Google Code Jam 2018 Round 1B
Round 1B was very difficult compared to 1A and 1C. I made some form of
mistake on all of the problems.

Rounding Error:
I had mostly the correct intuition but failed in the implementation.

    * My first issue was just being very slow at thinking about how one
      would properly 'bump' a number up by half a percentage point.
      Integer arithmetic has never been my strong suit, because it is
      often so inelegant. I need to work on that more.

    * My second mistake was failing to see that we should leave
      rounded-up numbers alone. I.e. going from a rounded-up number to
      another rounded-up number does not actually help the solution. Not
      only was my initial solution incorrect, but it was made more
      complicated by having to take this case into account. Again, I
      need better focus and better intuition with integer arithmetic.

Mysterious Road Signs:
The solution is actually quite simple if you believe that you can do it
in O(N) time. Deriving one optimal solution from the previous one is
rather straightforward. However, I was scared by the 10^5 limit and
thought that it must have been at least O(N log N).

I did also find another O(N log^2 N) solution involving binary searching
over a segment tree. (Simply store valid solutions in the nodes of the
segment tree and for each sign, binary search for the longest valid
solution starting at that sign.) It took me a long time to find this
solution because I was trying to be clever about how to check solution
validity, when the straightforward approach actually worked.

Transmutation:
Transmutation was a hard problem. Firstly, it took me a long time to
realise that:

    * We should always use the current best formula until it is
      exhausted.

    * We should stop once a metal depends on itself.

The first one is a matter of intuition. I got too concerned with random
edge cases to see the broader pattern. I did eventually come up with the
two observations, but I didn't properly implement cycle-checking. This
is because I'd forgotten about topological sorting, and I'd simply
checked that we couldn't visit another visited node (which is not
correct).