TCO 2018 Round 1B (StablePairsDiv1, ThreeSameLetters)
StablePairsDiv1:
I came up with the correct solution for the general case (maximise
distance between pairs and minimise distance within a pair), but I
messed up the edge case where we only need to generate one pair. I need
to be less hasty and more careful with checking edge cases.

ThreeSameLetters:
This was a straightforward dynamic program. As before, I messed up one
edge case and set

    nxt[j][k][1] = cur[i][j][0]

when I should have added instead. However, I caught this in testing. The
funny thing is that there is actually an O(L) dynamic program based on
exploiting the symmetries of the problem. It seems obvious in
retrospect, but it wasn't required, so I didn't think about it.