VK Cup 2018 Round 3 (Aztec Catacombs)
This was a fun problem. I immediately got the intuition that if you can
return to a given cave, then you can go directly to the exit.
Unfortunately, it took me longer to see that if there is a cave distance
2 away from the start, then you always have a path of length 4. The
deeper intuition here is that the corridor-toggling essentially
trivialises a lot of the pathfinding.

I missed one major case where you can have a path of length 5. It had
not occurred to me that this case could happen, and it was hard to prove
that length 5 was possible, but length 6 was not. Perhaps the lesson
here is that I need to practise constructing (or generating)
counterfactual cases.

The other slip-up was that once I knew about the case, I spent a long
time trying to actually figure out how to compute it -- it seemed I
needed to find a non-clique triple, and the dumb recursive approach
could be very slow. The solution is to find the node with smallest
degree and see if there is a node distance 2 away from it. It seems my
graph theory intuition is a bit rusty also.