Lyft Level 5 Challenge 2018 Final Round
Very minor change to segment trees. As a general observation, I think
I am pretty rusty and need to practise more. Somehow I'm making a lot
more blunders than I used to.

Таксисты и Lyft:
Wasted a few minutes on a silly overflow bug.

Ладья возвращается домой:
For some reason I kept on wanting to mentally model the problem as an
N^2 grid, which is not the right way to do this. It took me way too long
to figure out the correct solution (seeing that there are only a small
number of choices for vertical barriers).

Пересекающиеся поддеревья:
The intuition here was just to find the closest point to any point from
the unknown subtree. I'm honestly not sure what I was thinking.

Оптимальный периметр многоугольника:
With this problem, I really needed to just spend a bit more time on the
casework. It immediately saw that 4 and above was trivial -- it should
have not taken that much more work to see that with 3, we always wanted
to pick at least two extreme points.

Запросы на дедукцию:
This was pretty fun. What would have helped would have been to think
about partial sums earlier.

Поиск в глубину:
I got all the relevant observations, except for how to actually
construct the segment tree (min value, min count, delta). I was trying
to see if I could use (count zero, delta), but that doesn't work if you
have a -1 node above a 1 node. I ran out of ideas after that. I think
the main point of intuition was seeing that min and min count played
better with delta than count zero did, since it associates better.