TopCoder SRM 746
Python was surprisingly good for this SRM because there were not
actually any computationally complex problems.

ChangeDistances:
Easy-ish problem. Just notice that there is an edge between two vertices
if and only if the two vertices are distance one apart (by definition).

FindStringHard:
I got an initial solution pretty quickly, but it turned out to be wrong.
I didn't completely check things because I felt like I was in a hurry,
so the lesson here is to be a bit less anxious and feel less pressured
by time.

SpaceProbes:
This was a funny problem. I completely failed reading comprehension and
thought this problem took place between two line segments (which made
the problem much, much harder). This is especially embarrassing, since
I'm trying to write something right now. Hah.

After correctly reading the problem statement, the math itself was quite
simple, but I had a few issues with precision. I missed a bug where
dividing by something accidentally converted a number into a double
(from a big integer), leading to a loss in precision. So the trick to
remember was, perhaps, to keep everything as big integers until the very
end. But in the end, I went with the cheap way out and decided to use the
`decimal` package instead. This might be a crutch in the long run, haha.

So aside from failing at reading comprehension and implementation, the
math/algorithm aspect of the problem was quite easy. This is a bit
abnormal for a 1000-pointer.